rename url options
